Improve MessageIO -> IRC conversion
