--- Sources/D2IRCIO/IRCChannelNameConverter.swift
@@ -1,10 +1,0 @@
-import D2MessageIO
-import IRC
-
-// TO IRC conversions
-
-extension ID: IRCAPIConvertible {
-    public var usingIRCAPI: IRCChannelName {
-        IRCChannelName(value)!
-    }
-}
