+++ Sources/D2IRCIO/IRCMessageClient.swift
@@ -1,4 +1,5 @@
 import D2MessageIO
+import Emoji
 import IRC
 import Logging
 
@@ -51,10 +52,33 @@ struct IRCMessageClient: MessageClient {
         then?(nil, nil)
     }
 
+    private func flatten(embed: Embed) -> String {
+        let lines: [String?] = [
+            embed.title.flatMap { title in embed.url.map { "[\(title)](\($0.absoluteString))" } ?? title },
+            embed.description
+        ] + embed.fields.flatMap { ["**\($0.name)**", $0.value] } + [
+            embed.footer?.text
+        ]
+        return lines
+            .compactMap { $0 }
+            .joined(separator: "\n")
+    }
 	
 	func sendMessage(_ message: D2MessageIO.Message, to channelId: ChannelID, then: ClientCallback<D2MessageIO.Message?>?) {
         log.debug("Sending message '\(message.content)'")
-        ircClient.sendMessage(message.usingIRCAPI)
+
+        let text = [message.content, message.embed.map(flatten(embed:))]
+            .compactMap { $0?.nilIfEmpty }
+            .joined(separator: "\n")
+            .emojiUnescapedString
+
+        guard let channelName = IRCChannelName(channelId.value) else {
+            log.warning("Could not convert \(channelId.value) (maybe it is missing a leading '#'?)")
+            return
+        }
+
+        ircClient.send(.PRIVMSG([.channel(channelName)], text))
+        // TODO: Handle client callback here
     }
 	
 	func editMessage(_ id: MessageID, on channelId: ChannelID, content: String, then: ClientCallback<D2MessageIO.Message?>?) {
