--- Sources/D2IRCIO/IRCMessageConverter.swift
@@ -1,28 +1,0 @@
-import D2MessageIO
-import Emoji
-import IRC
-
-// TO IRC conversions
-    
-private func flatten(embed: Embed) -> String {
-    let lines: [String?] = [
-        embed.title.flatMap { title in embed.url.map { "[\(title)](\($0.absoluteString))" } ?? title },
-        embed.description
-    ] + embed.fields.flatMap { ["**\($0.name)**", $0.value] } + [
-        embed.footer?.text
-    ]
-    return lines
-        .compactMap { $0 }
-        .joined(separator: "\n")
-}
-
-extension D2MessageIO.Message: IRCAPIConvertible {
-    public var usingIRCAPI: IRCMessage {
-        let text = [content, embed.map(flatten(embed:))]
-            .compactMap { $0?.nilIfEmpty }
-            .joined(separator: "\n")
-            .emojiUnescapedString
-        // TODO: Find a better solution than force-unwrapping
-        return IRCMessage(command: .PRIVMSG(channelId.map { [.channel(IRCChannelName($0.value)!)] } ?? [], text))
-    }
-}
